now we save users to database

+++ internal/database/postgres/gorm/coordinates/methods.go
@@ -12,11 +12,11 @@ func (c *coordinates) CreateTableIfNotExists() error {
 	return nil
 }
 
-func (c *coordinates) Select(userId uint) (model.Coordinates, error) {
+func (c *coordinates) Select(userId uint64) (model.Coordinates, error) {
 	return model.Coordinates{}, errors.New("method not implemented")
 }
 
-func (c *coordinates) SelectMany(userId, limit, offset uint) ([]model.Coordinates, error) {
+func (c *coordinates) SelectMany(userId, limit, offset uint64) ([]model.Coordinates, error) {
 	return nil, errors.New("method not implemented")
 }
 
@@ -24,6 +24,6 @@ func (c *coordinates) Insert(coordinates model.Coordinates) (model.Coordinates, 
 	return model.Coordinates{}, errors.New("method not implemented")
 }
 
-func (c *coordinates) Delete(userId uint) error {
+func (c *coordinates) Delete(userId uint64) error {
 	return errors.New("method not implemented")
 }
